requirements.txt
streamlit==1.21.0
googletrans==4.0.0-rc1
pandas==1.5.3
matplotlib==3.7.1
python-Levenshtein==0.12.2
